requirements.txt
fastapi[standard]==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
python-gitlab==6.3.0
jira==3.10.5
atlassian-python-api==4.0.7
python-multipart==0.0.6
pydantic==2.7.0
python-dotenv==1.0.0
jinja2==3.1.2
